+++ src/platform/linux.ts
@@ -36,7 +36,7 @@ export class LinuxPortProvider implements InspectPortProvider {
     };
   }
 
-  async killProcesses(processes: ProcessInfo[]): Promise<void> {
-    await killWithSignal(processes);
+  async killProcesses(processes: ProcessInfo[], signal?: NodeJS.Signals): Promise<void> {
+    await killWithSignal(processes, signal);
   }
 }